+++ src/frontend/apps/impress/src/features/docs/doc-editor/types.tsx
@@ -1,3 +1,5 @@
+import { BlockNoteEditor, BlockNoteSchema } from '@blocknote/core';
+
 export interface DocAttachment {
   file: string;
 }
@@ -12,3 +14,19 @@ export type HeadingBlock = {
     level: number;
   };
 };
+
+export const blockNoteInstance = BlockNoteSchema.create();
+export type DocsBlockSchema = typeof blockNoteInstance.blockSchema;
+export type DocsInlineContentSchema =
+  typeof blockNoteInstance.inlineContentSchema;
+export type DocsStyleSchema = typeof blockNoteInstance.styleSchema;
+export type DocsBlockNoteEditor = BlockNoteEditor<
+  DocsBlockSchema,
+  DocsInlineContentSchema,
+  DocsStyleSchema
+>;
+export type DocsBlockNoteSchema = BlockNoteSchema<
+  DocsBlockSchema,
+  DocsInlineContentSchema,
+  DocsStyleSchema
+>;