+++ CHANGELOG.md
@@ -10,7 +10,13 @@ and this project adheres to
 ## [Unreleased]
 
 ## Added
+
 - 📝(doc) Add security.md and codeofconduct.md #604
+
+## Fixed
+
+- 💄improve export spacings PDF #613
+
 
 ## [2.1.0] - 2025-01-29